+++ packages/store/src/StoreReadWithStubs.sol
@@ -59,7 +59,7 @@ contract StoreReadWithStubs is IStore, StoreRead {
   /**
    * Not implemented in StoreReadWithStubs
    */
-  function registerStoreHook(bytes32, IStoreHook) public virtual {
+  function registerStoreHook(bytes32, IStoreHook, uint8) public virtual {
     revert StoreReadWithStubs_NotImplemented();
   }