+++ packages/protocol-parser/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @latticexyz/protocol-parser
 
+## 2.0.0-next.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @latticexyz/common@2.0.0-next.7
+  - @latticexyz/schema-type@2.0.0-next.7
+
+## 2.0.0-next.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @latticexyz/schema-type@2.0.0-next.6
+  - @latticexyz/common@2.0.0-next.6
+
 ## 2.0.0-next.5
 
 ### Patch Changes