+++ packages/store-sync/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @latticexyz/store-sync
 
+## 2.0.0-next.7
+
+### Patch Changes
+
+- Updated dependencies [[`c4d5eb4e`](https://github.com/latticexyz/mud/commit/c4d5eb4e4e4737112b981a795a9c347e3578cb15), [`18d3aea5`](https://github.com/latticexyz/mud/commit/18d3aea55b1d7f4b442c21343795c299a56fc481)]:
+  - @latticexyz/store@2.0.0-next.7
+  - @latticexyz/world@2.0.0-next.7
+  - @latticexyz/block-logs-stream@2.0.0-next.7
+  - @latticexyz/common@2.0.0-next.7
+  - @latticexyz/protocol-parser@2.0.0-next.7
+  - @latticexyz/recs@2.0.0-next.7
+  - @latticexyz/schema-type@2.0.0-next.7
+
+## 2.0.0-next.6
+
+### Patch Changes
+
+- Updated dependencies [[`8025c350`](https://github.com/latticexyz/mud/commit/8025c3505a7411d8539b1cfd72265aed27e04561)]:
+  - @latticexyz/store@2.0.0-next.6
+  - @latticexyz/world@2.0.0-next.6
+  - @latticexyz/schema-type@2.0.0-next.6
+  - @latticexyz/block-logs-stream@2.0.0-next.6
+  - @latticexyz/common@2.0.0-next.6
+  - @latticexyz/protocol-parser@2.0.0-next.6
+  - @latticexyz/recs@2.0.0-next.6
+
 ## 2.0.0-next.5
 
 ### Patch Changes