+++ packages/store/CHANGELOG.md
@@ -1,5 +1,109 @@
 # Change Log
 
+## 2.0.0-next.7
+
+### Major Changes
+
+- [#1399](https://github.com/latticexyz/mud/pull/1399) [`c4d5eb4e`](https://github.com/latticexyz/mud/commit/c4d5eb4e4e4737112b981a795a9c347e3578cb15) Thanks [@alvrs](https://github.com/alvrs)! - - The `onSetRecord` hook is split into `onBeforeSetRecord` and `onAfterSetRecord` and the `onDeleteRecord` hook is split into `onBeforeDeleteRecord` and `onAfterDeleteRecord`.
+  The purpose of this change is to allow more fine-grained control over the point in the lifecycle at which hooks are executed.
+
+  The previous hooks were executed before modifying data, so they can be replaced with the respective `onBefore` hooks.
+
+  ```diff
+  - function onSetRecord(
+  + function onBeforeSetRecord(
+      bytes32 table,
+      bytes32[] memory key,
+      bytes memory data,
+      Schema valueSchema
+    ) public;
+
+  - function onDeleteRecord(
+  + function onBeforeDeleteRecord(
+      bytes32 table,
+      bytes32[] memory key,
+      Schema valueSchema
+    ) public;
+  ```
+
+  - It is now possible to specify which methods of a hook contract should be called when registering a hook. The purpose of this change is to save gas by avoiding to call no-op hook methods.
+
+    ```diff
+    function registerStoreHook(
+      bytes32 tableId,
+    - IStoreHook hookAddress
+    + IStoreHook hookAddress,
+    + uint8 enabledHooksBitmap
+    ) public;
+
+    function registerSystemHook(
+      bytes32 systemId,
+    - ISystemHook hookAddress
+    + ISystemHook hookAddress,
+    + uint8 enabledHooksBitmap
+    ) public;
+    ```
+
+    There are `StoreHookLib` and `SystemHookLib` with helper functions to encode the bitmap of enabled hooks.
+
+    ```solidity
+    import { StoreHookLib } from "@latticexyz/store/src/StoreHook.sol";
+
+    uint8 storeHookBitmap = StoreBookLib.encodeBitmap({
+      onBeforeSetRecord: true,
+      onAfterSetRecord: true,
+      onBeforeSetField: true,
+      onAfterSetField: true,
+      onBeforeDeleteRecord: true,
+      onAfterDeleteRecord: true
+    });
+    ```
+
+    ```solidity
+    import { SystemHookLib } from "@latticexyz/world/src/SystemHook.sol";
+
+    uint8 systemHookBitmap = SystemHookLib.encodeBitmap({
+      onBeforeCallSystem: true,
+      onAfterCallSystem: true
+    });
+    ```
+
+  - The `onSetRecord` hook call for `emitEphemeralRecord` has been removed to save gas and to more clearly distinguish ephemeral tables as offchain tables.
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @latticexyz/common@2.0.0-next.7
+  - @latticexyz/config@2.0.0-next.7
+  - @latticexyz/schema-type@2.0.0-next.7
+
+## 2.0.0-next.6
+
+### Minor Changes
+
+- [#1413](https://github.com/latticexyz/mud/pull/1413) [`8025c350`](https://github.com/latticexyz/mud/commit/8025c3505a7411d8539b1cfd72265aed27e04561) Thanks [@holic](https://github.com/holic)! - We now use `@latticexyz/abi-ts` to generate TS type declaration files (`.d.ts`) for each ABI JSON file. This replaces our usage TypeChain everywhere.
+
+  If you previously relied on TypeChain types from `@latticexyz/store` or `@latticexyz/world`, you will either need to migrate to viem or abitype using ABI JSON imports or generate TypeChain types from our exported ABI JSON files.
+
+  ```ts
+  import { getContract } from "viem";
+  import IStoreAbi from "@latticexyz/store/abi/IStore.sol/IStore.abi.json";
+
+  const storeContract = getContract({
+    abi: IStoreAbi,
+    ...
+  });
+
+  await storeContract.write.setRecord(...);
+  ```
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @latticexyz/schema-type@2.0.0-next.6
+  - @latticexyz/common@2.0.0-next.6
+  - @latticexyz/config@2.0.0-next.6
+
 ## 2.0.0-next.5
 
 ### Patch Changes